Fix tests for useDeleteComment
