+++ x-pack/plugins/cases/public/components/case_view/__mocks__/use_on_refresh_case_view_page.tsx
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+const refreshFunction = jest.fn();
+export const useRefreshCaseViewPage = () => refreshFunction;
