+++ x-pack/plugins/cases/public/containers/use_delete_comment.test.tsx
@@ -9,16 +9,13 @@ import { act, renderHook } from '@testing-library/react-hooks';
 import { useDeleteComment, UseDeleteComment } from './use_delete_comment';
 import * as api from './api';
 import { basicCaseId } from './mock';
-import { useRefreshCaseViewPage } from '../components/case_view/case_view_page';
+import { useRefreshCaseViewPage } from '../components/case_view/use_on_refresh_case_view_page';
 
 jest.mock('../common/lib/kibana');
 jest.mock('./api');
-jest.mock('../components/case_view/case_view_page');
+jest.mock('../components/case_view/use_on_refresh_case_view_page');
 
 const commentId = 'ab124';
-
-const refreshMock = jest.fn();
-(useRefreshCaseViewPage as jest.Mock).mockReturnValue(refreshMock);
 
 describe('useDeleteComment', () => {
   it('init', async () => {
@@ -69,7 +66,7 @@ describe('useDeleteComment', () => {
         commentId,
       });
       await waitForNextUpdate();
-      expect(refreshMock).toBeCalled();
+      expect(useRefreshCaseViewPage()).toBeCalled();
     });
   });
